Extract evm
